[Deps] update to Diem 1.6.2 (#31)
### Merging Updates from diem-1.6.2 into main

This PR consolidates the advancements from the upstream diem repository and the work completed on 0D's libra-framework via the diem-1.6.2 branch. After thorough review and testing, these changes are now ready to be integrated into the main branch for the 0L network. This merge will squash the commits for a cleaner history.